Cleanup hlo_extractor and hlo_bisect dependecies
//xla/tools:hlo_extractor and //xla/tools/hlo_bisect:hlo_bisect_state
depend on //xla/tests:test_utils. This dependency should not exist.
Moving the relevant functions to //xla:literal_util.
