+++ xla/tools/hlo_bisect/BUILD
@@ -46,7 +46,7 @@ cc_library(
         "//xla:util",
         "//xla/hlo/ir:hlo",
         "//xla/service:hlo_dce",
-        "//xla/tests:test_utils",
+        "//xla:literal_util",
         "@com_google_absl//absl/algorithm:container",
         "@com_google_absl//absl/container:flat_hash_map",
         "@com_google_absl//absl/container:flat_hash_set",
