+++ xla/tools/hlo_bisect/hlo_bisect_state.cc
@@ -29,7 +29,7 @@ limitations under the License.
 #include "xla/hlo/ir/hlo_module.h"
 #include "xla/hlo/ir/hlo_opcode.h"
 #include "xla/service/hlo_dce.h"
-#include "xla/tests/test_utils.h"
+#include "xla/literal_util.h"
 #include "xla/util.h"
 
 namespace xla {
