+++ xla/tools/hlo_extractor.cc
@@ -42,7 +42,6 @@ limitations under the License.
 #include "xla/service/hlo_verifier.h"
 #include "xla/shape.h"
 #include "xla/shape_util.h"
-#include "xla/tests/test_utils.h"
 #include "tsl/platform/status.h"
 
 namespace xla {
